add trace param
